Wrap up home display with generated cards. Create configure page.
