+++ App/src/assets/main.css
@@ -4,31 +4,19 @@
   max-width: 1280px;
   margin: 0 auto;
   font-weight: normal;
+  padding: 0 2rem;
 }
 
 a,
-.green {
+.blue {
   text-decoration: none;
-  color: hsla(160, 100%, 37%, 1);
+  color: rgb(7, 170, 235);
   transition: 0.4s;
   padding: 3px;
 }
 
 @media (hover: hover) {
   a:hover {
-    background-color: hsla(160, 100%, 37%, 0.2);
-  }
-}
-
-@media (min-width: 1024px) {
-  body {
-    display: flex;
-    place-items: center;
-  }
-
-  #app {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    padding: 0 2rem;
+    background-color: hsla(214, 100%, 37%, 0.2);
   }
 }
