--- App/src/components/WelcomeItem.vue
@@ -1,28 +1,0 @@
-<template>
-  <div class="card">
-    <div class="label">
-
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.card {
-  background-color: #ccc;
-  display: inline-block;
-  border-color: #333;
-  border-radius: 10%;
-  width: 2.5rem;
-  height: 3rem;
-}
-
-.label {
-  font-size: 1.5rem;
-  font-weight: 900;
-  color: #333;
-}
-
-@media (min-width: 1024px) {
-
-}
-</style>
