+++ App/src/router/index.ts
@@ -1,22 +1,21 @@
 import { createRouter, createWebHistory } from 'vue-router'
+import HomeView from '@/views/HomeView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
-    /*
     {
       path: '/',
       name: 'home',
       component: HomeView
     },
-    */
     {
-      path: '/about',
-      name: 'about',
+      path: '/configure',
+      name: 'configure',
       // route level code-splitting
-      // this generates a separate chunk (About.[hash].js) for this route
+      // this generates a separate chunk (Configure.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
-      component: () => import('../views/AboutView.vue')
+      component: () => import('../views/ConfigureView.vue')
     }
   ]
 })
